+++ sickbeard/providers/freshontv.py
@@ -101,6 +101,7 @@ class FreshOnTVProvider(generic.TorrentProvider):
                         if 2 > len(torrent_rows):
                             raise generic.HaltParseException
 
+                        head = None
                         for tr in torrent_rows[1:]:
                             cells = tr.find_all('td')
                             if (5 > len(cells) or tr.find('img', alt='Nuked')
@@ -109,8 +110,9 @@ class FreshOnTVProvider(generic.TorrentProvider):
                                          or (not non_marked and not tr.find('img', src=rc['filter']))))):
                                 continue
                             try:
+                                head = head if None is not head else self._header_row(tr)
                                 seeders, leechers, size = [tryInt(n, n) for n in [
-                                    cells[x].get_text().strip() for x in -2, -1, -4]]
+                                    cells[head[x]].get_text().strip() for x in 'seed', 'leech', 'size']]
                                 if self._peers_fail(mode, seeders, leechers):
                                     continue