+++ sickbeard/providers/gftracker.py
@@ -81,12 +81,14 @@ class GFTrackerProvider(generic.TorrentProvider):
                         if 2 > len(torrent_rows):
                             raise generic.HaltParseException
 
+                        head = None
                         for tr in torrent_rows[1:]:
                             cells = tr.find_all('td')
                             if 3 > len(cells):
                                 continue
                             try:
-                                seeders, leechers = 2 * [cells[-1].get_text().strip()]
+                                head = head if None is not head else self._header_row(tr)
+                                seeders, leechers = 2 * [cells[head['seed']].get_text().strip()]
                                 seeders, leechers = [tryInt(n) for n in [
                                     rc['seeders'].findall(seeders)[0], rc['leechers'].findall(leechers)[0]]]
                                 if self._peers_fail(mode, seeders, leechers):
@@ -94,7 +96,7 @@ class GFTrackerProvider(generic.TorrentProvider):
 
                                 info = tr.find('a', href=rc['info'])
                                 title = (info.attrs.get('title') or info.get_text()).strip()
-                                size = cells[-2].get_text().strip()
+                                size = cells[head['size']].get_text().strip()
                                 download_url = self._link(tr.find('a', href=rc['get'])['href'])
                             except (AttributeError, TypeError, ValueError):
                                 continue