+++ sickbeard/providers/hdtorrents.py
@@ -106,6 +106,7 @@ class HDTorrentsProvider(generic.TorrentProvider):
                         if not len(torrent_rows):
                             raise generic.HaltParseException
 
+                        head = None
                         for tr in torrent_rows:
                             cells = tr.find_all('td')
                             if (6 > len(cells) or any(self.filter)
@@ -113,8 +114,9 @@ class HDTorrentsProvider(generic.TorrentProvider):
                                      or (not non_marked and not tr.find('img', src=rc['filter'])))):
                                 continue
                             try:
+                                head = head if None is not head else self._header_row(tr)
                                 seeders, leechers, size = [tryInt(n, n) for n in [
-                                    cells[x].get_text().strip() for x in -3, -2, -5]]
+                                    cells[head[x]].get_text().strip() for x in 'seed', 'leech', 'size']]
                                 if self._peers_fail(mode, seeders, leechers) or not tr.find('a', href=rc['cats']):
                                     continue
                                 title = tr.find('a', href=rc['info']).get_text().strip()