+++ sickbeard/providers/torrentbytes.py
@@ -74,14 +74,16 @@ class TorrentBytesProvider(generic.TorrentProvider):
                         if 2 > len(torrent_rows):
                             raise generic.HaltParseException
 
+                        head = None
                         for tr in torrent_rows[1:]:
                             cells = tr.find_all('td')
                             if 5 > len(cells):
                                 continue
                             try:
                                 info = tr.find('a', href=rc['info'])
+                                head = head if None is not head else self._header_row(tr)
                                 seeders, leechers, size = [tryInt(n, n) for n in [
-                                    cells[x].get_text().strip() for x in -2, -1, -4]]
+                                    cells[head[x]].get_text().strip() for x in 'seed', 'leech', 'size']]
                                 if self.freeleech and (len(info.contents) < 2 or not rc['fl'].search(
                                         info.contents[1].string.strip())) or self._peers_fail(mode, seeders, leechers):
                                     continue