+++ CHANGES.md
@@ -223,6 +223,7 @@
 * Fix PiSexy for vip user class
 * Change retry sleep timeout for Trakt searches to prevent issues when Trakt is down
 * Fix TVDb search issue when only 1 result is returned
+* Change improve TvChaos item parsing and can use qualities instead of 'Unknown'
 
 
 ### 0.11.16 (2016-10-16 17:30:00 UTC)